// src/enums/Route.ts
export enum Route {
  SIGN_IN = 'SignIn',
  SIGN_UP = 'SignUp',
  RESET_PASSWORD = 'ResetPassword',

  FINANCES = 'Finances',
  FINANCE_MANAGER = 'FinanceManager',
  ANALYTICS = 'Analytics',
  CATEGORIES = 'Categories',
  CATEGORY = 'Category',
  CATEGORY_MANAGER = 'CategoryManager',
  SEARCH = 'Search',
  SETTINGS = 'Settings',

  PROFILE = 'Profile',
  USER = 'User',
  USERS_CONNECT = 'UsersConnect',
  BANKS_CONNECT = 'BanksConnect',
  PAYMENT_METHODS = 'PaymentMethods',
  LOCAL_AUTHENTICATION = 'LocalAuthentication',
  APP_ICON = 'AppIcon',
  APPEARANCE = 'Appearance',
  BOTTOM_NAV_ITEMS = 'BottomNavItems',

  MAIN = 'Main',

  LOGGED_IN = 'LoggedIn',
  LOGGED_OUT = 'LoggedOut',
}